Initial work to convert individual power readings to a second.
A single `DataEntry` represents a single power reading in time. We want to work on a data entry that represents power over a second.
To do this we need to represent a seconds worth of power as a point of time at the start of a second and a point of time at the end of the second. Currently this is represented as a `SecondPower`.