Add `Plottable` interface and `Ride`.
A `Ride` will contain data  about a ride, e.g. Power, Heart Rate & Cadence (in the future speed and elevation). Each of those each of those pieces of data are stored in an independent array. This means if power is missing for the time 10:00:00 but we have Cadence and Heart Rate information for that time it can still be recorded and plotted. The `DataEntry` class now contains just a time and a value.

The `Plottable` interface will define a way to provide a `Map<String, DataEntry[]>` that can be plotted onto a graph. For the `Ride` class Power, Cadence and Heart Rate are all plottable so a `Map` providing that data will be provided.

Potential idea to replace the `Map` with a `Plot` class providing more information.